requirements.txt
praw
python-dotenv
requests
beautifulsoup4
lxml